Uncap meso balances to 9,999,999,999,999
- GameConstants: MAX_MESO = 9,999,999,999,999.
- Character: wallet and merchant meso are long; canHoldMeso, gainMeso
  and merchant withdrawal cap at MAX_MESO; load/save as long.
- Storage: meso is long; the storage meso transfer caps at MAX_MESO.
- PacketCreator: the wallet (character info, STAT_CHANGED), Storage,
  Fredrick and Hired Merchant balances are written as long, matching
  the Kaentake maxmeso client mod. Single-transaction amounts (trade,
  prices, RPS, drops) stay int.
- !givems accepts amounts up to MAX_MESO.
- DB: changeset 26 widens characters.meso, characters.MerchantMesos and
  storages.meso to BIGINT.

+++ src/main/java/net/server/channel/handlers/RangedAttackHandler.java
@@ -135,7 +135,7 @@ public final class RangedAttackHandler extends AbstractDealDamageHandler {
                         int moneyMod = money / 2;
                         money += Randomizer.nextInt(moneyMod);
                         if (money > chr.getMeso()) {
-                            money = chr.getMeso();
+                            money = (int) chr.getMeso(); // below money, so it fits an int
                         }
                         chr.gainMeso(-money, false);
                     }
